renamed pages to views
