+++ app/views/Logout.js
